banner_web_app/requirements.txt
Flask==1.1.2
gunicorn>=19,<20
redis==3.4.1
mysql-connector-python